SDL3 by default (#933)
* Windows: bundle SDL3 3.4.12 (x64 headers, VC and MinGW import libs)

* Finish optional SDL3 support: MSVC configs, makefiles, CI and packaging

- MSVC: new Debug-SDL3/Release-SDL3 configurations using a derived
  BaseConfiguration property; SDL include/lib dirs, import libs, USE_SDL3
  define and post-build DLL copy are parametrized. Helper projects map to
  the plain Debug/Release configs in the solution.
- GNU makefiles: USE_SDL3=1 knob on Linux, MinGW x64 (bundled SDL3) and
  MSYS2 clangarm64 (pkg-config). in_sdl*/snd_sdl* backends are now
  self-guarding and always compiled; selection is by define only.
- main_sdl.c includes SDL3/SDL_main.h (header-only WinMain shim, SDL3 has
  no SDL3main library).
- CI: SDL3 legs for MSVC, mingw64, clangarm64 and macOS; AppImage docker
  image builds SDL3 from source and compile-checks the SDL3 backend.
- Fix SDL3 backend issues found by MSVC /W4: TASK_TIMEOUT_INFINITE
  signedness, invalid SDL_JoystickID is 0 in SDL3, gamepad events use
  gdevice (cdevice is the camera event in SDL3).
- snd_sdl3.c: use SDL_Pause/ResumeAudioStreamDevice, log the device name.

* Make SDL3 the default in all build systems

- MSVC: plain Debug/Release configurations now build against the bundled
  SDL3; the separate Debug-SDL3/Release-SDL3 configurations are gone.
- Meson: use_sdl3 is a feature option (default auto). SDL3 is preferred
  and the build falls back to SDL2 when it is not installed; disabled
  forces SDL2, enabled requires SDL3.
- GNU makefiles: USE_SDL3 defaults to 1; pass USE_SDL3=0 for SDL2.
- AppImage now ships SDL3 (built from source in the docker image,
  ldconfig so linuxdeploy can resolve it); SDL2 backend is compile-checked.
- CI: default legs build SDL3 everywhere, one SDL2 leg per platform keeps
  the SDL2 code paths building.

* Makefile.w64: fix stale comment, SDL2 is the fallback not the default

* Remove SDL2 from Windows builds entirely

Windows always has the bundled SDL3, so the vendored Windows/SDL2 files,
the USE_SDL3 knob in the MinGW makefiles and the SDL2 CI legs are gone.
SDL2 support remains for Linux source builds only (meson auto-fallback,
USE_SDL3=0 in the unix Makefile, AppImage compile check).

* Remove dead SDL2 code paths from Windows-only code

pl_win.c and the VK_EXT_full_screen_exclusive block in gl_vidsdl.c are
only ever compiled on Windows, which is SDL3 only now; drop the
SDL_syswm.h based HWND lookups.

* Fix SDL3 video/window issues

- VID_SDL_GetDisplayMode: fix use-after-free. SDL3's
  SDL_GetFullscreenDisplayModes returns a single allocation that contains
  the mode structs themselves, so returning a pointer into it after
  SDL_free read freed memory (intermittent wrong/failed fullscreen modes).
  The best match is now copied out before freeing.
- Refresh rates are kept as exact floats end to end (mode list, cvar,
  menu) instead of being truncated to int, and mode matching picks the
  mode with the closest refresh rate when there is no exact match, so
  configs written by SDL2 builds or other displays still resolve.
- Mode lookups use the display the window is on instead of always the
  primary display (both SDL2 and SDL3 paths).
- VID_SetMode: SDL_SyncWindow after showing the window; SDL3 window size,
  position and fullscreen changes are async requests on X11/Wayland and
  the sizes were queried before they had been applied.
- in_sdl3.c: handle SDL_EVENT_WINDOW_PIXEL_SIZE_CHANGED instead of
  RESIZED; data1/data2 are pixels (matching vid.width/height) and it also
  fires when only the display scale changes.

* Tighten SDL3 comments to facts the code can't express

* Accumulate mouse motion as float, SDL3 reports subpixel relative motion

Truncating SDL3's float xrel/yrel per event loses fractional motion with
a bias toward zero, which shows up as notchy low speed aim. Also remove
an unused SDL2 event shim and duplicated defines.

* clang-format

* mathlib.h: include arm_neon.h for the NEON code paths

SDL2's SDL_cpuinfo.h pulled in arm_neon.h transitively; SDL3 moved the
intrinsics includes to SDL_intrin.h, which SDL.h does not include. The
SSE2 branch already includes emmintrin.h explicitly.

* gl_vidsdl.c: include windows.h explicitly for vulkan_win32.h

The Win32 types came in transitively through SDL2's SDL_syswm.h, which is
no longer included; MinGW has no other include path that provides them.

* steam.c: SDL_ShowMessageBox returns bool on SDL3

* AppImage docker: add X11 dev packages required by SDL3

SDL3's cmake hard-requires XInput2, Xext, Xfixes for the X11 driver
(SDL2's configure silently disabled missing ones). Also add Xss for the
screensaver inhibit support.

* AppImage docker: add libxtst-dev, completes SDL3's required X11 set

Cross-checked against every fatal SDL_missing_dependency in SDL 3.4.12's
CheckX11: XCURSOR, XDBE, XINPUT, XFIXES, XSYNC, XRANDR, XSCRNSAVER,
XSHAPE and XTEST are now all satisfied.

* Makefile.w64a: don't use pkg-config for SDL3

pkg-config isn't reliably available in the clangarm64 CI environment and
the $(shell) failure is silent, producing a link line without -lSDL3.
Hardcode the flags like the rest of the makefile does; the headers and
import library come from the MSYS2 prefix that is already on the include
and library paths.

+++ Quake/tasks.h
@@ -80,7 +80,8 @@ void TestTasks_f (void);
 #endif
 
 #ifdef USE_SDL3
-#define TASK_TIMEOUT_INFINITE -1
+// passed through to SDL3's Sint32 millisecond timeouts, where -1 means infinite
+#define TASK_TIMEOUT_INFINITE ((uint32_t) - 1)
 #else
 #define TASK_TIMEOUT_INFINITE SDL_MUTEX_MAXWAIT
 #endif
